created Role class for enum, role column in user, new constructor in Customer for must have data,Autentication with Jwt Tokken added and work,

+++ src/main/java/com/KacFlor/ShopSpring/dao/UserRepository.java
@@ -1,12 +1,14 @@
 package com.KacFlor.ShopSpring.dao;
 
 import com.KacFlor.ShopSpring.model.User;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer>{
+
     User findByLogin(String login);
 }
